Remove service instances, add labels to service (#83)
After much discussion, we decided that we will keep
global labels as they are for now, and eventually
remove them when we have a replacement. The replacement
is expected to be something along the lines of more
explicit, user-configurable dimensions. Between now
and then we may add more fields to the service
aggregation key for common use cases, e.g. service
namespace.

The service instances concept does not align with
that future, so we're reverting to aggregating by
global labels at the service level. The implication
is that global labels are expected to have low
cardinality, and if they do not then users can expect
a service with high cardinality global labels to
interfere with the collection of metrics for other
services (i.e. causing them to overflow.)