enhance sorting logic
TestAggregateAndHarvest sort logic only used Metricset.Name.
The latest test added to it s flappy with such a logic, as we are
evaluating the same Metricset.Names in 2 different events with
global labels with equal key and different values.

Enhance the sorting logic to account for equal Metricset.Name and
leverage labels (keys, Value and Values) in the lessFn.